requirements.txt
beautifulsoup4==4.12.3
requests == 2.32.3
lxml == 5.2.2
PyPDF2 ~= 2.0
python-dotenv == 1.0.1
aiohttp == 3.9.5
pandas == 2.2.2
flask == 3.0.3
PyYAML == 6.0.2
flask-cors == 5.0.0
rapidfuzz == 3.9.5
uvicorn == 0.30.6
Gunicorn == 23.0.0
polars == 1.8.2